Refactor logging to extract complex functionality from log formatters (#97)
## Changes

* Extract functionality of adding Flask request context data out of log
formatters
  into flask_logger module
* Extract PII masking functionality to pii module
* Move mask_pii and mask_pii_for_key from string_utils to pii module
* Use logging module main API for configuration instead of
logging.config.dictConfig
* Remove custom implementation of log capture fixture
* Import standard logging module instead of api.logging across
application
* Automatically log request at the beginning of each request
* Bring test coverage to 100% for updated modules

## Context

As I was looking at the logging code to document how it works (ticket
#87) and add the enhancement to add custom data to current request's
logs (ticket #95), I observed that a lot of functionality is mixed into
the formatter classes. In particular, the log formatters weren't just
doing formatting, but they were also (1) adding additional attributes
from the flask request context, and (2) masking PII. This makes the
logging system much harder to maintain, since complexity of the
formatters is so high as they are doing so many things. It also creates
inconsistencies and duplicate code. For example, the logic to add
request context data is duplicated in both the JSON and human readable
formatter, and would need to be duplicated for any new formatter. As
another example, PII masking is only done for the JSON formatter, not
the human readable formatter, which is an inconsistency that isn't
obvious from looking at the high level purpose of the two formatters.

This looks like a big PR but there's no new functionality in this PR. A
lot of the new lines are docstrings and comments, and the boilerplate
imports and whitespace you get when separating out functionality into
multiple files. The following summarizes the changes:

1. This PR pulls out the logic that isn't related to formatting out of
the log formatters, which simplifies the code to be more maintainable as
each module's functionality is now more restricted.

2. This PR also aims to reduce the number of generic "utils" by moving
mask_pii and mask_pii_for_key to the api.logging.pii module rather than
leaving it in a generic string_utils module.

3. As part of this refactoring, I also noticed a fix_logging pytest
fixture and a lot of code in test_logging that seems to emulate pytest's
caplog. I realized that it's because the way the application configures
logging uses logging.config.dictConfig which blows away any existing
logging handlers. In particular, it clears the caplog handler that the
caplog fixture adds to the root logger. This PR changes the
configuration to use the main logging API which involve non-destructive
functions like addHandler, which more cleanly ignores any existing
handlers. This is more appropriately follows the listener pattern and
also allows us to remove a lot of extra testing code.

4. Finally, I didn't really understand the purpose of forcing the
application to include api.logging instead of the standard module
logging, and to rename getLogger as get_logger. The benefit doesn't seem
to outweigh the cons of reducing the portability/reusability of code and
the necessity of learning a slightly different way of logging. Since
logging is a core part of Python, I refactored the application to use
the standard library, and only use api.logging for initialization and
configuration and custom enhanced behavior.

With this refactor, we are now more easily able to add additional
enhancements while keeping complexity low. In this PR we added the
ability to automatically log the request at the beginning of each flask
request, so we no longer need to do that in each route. And in a follow
up PR https://github.com/navapbc/template-application-flask/pull/96 we
are able to add a function that adds custom data that gets added to all
logs for the current flask request.