Move LRUDict to audit module (#135)
* Move LeastRecentlyUsedDict to audit module

## Context for reviewers
Moving the LeastRecentlyUsedDict to the audit module since it's only
used in one place, and the particular implementation of the dict is
specific to the audit module since it is designed to store integers (it
defaults values to `0`). We also want to minimize catch all utilities
packages/modules to have cleaner dependencies and better portability.