Rename application package to src from api (#146)
In PR #138 we renamed the route package to api, which makes it api.api.
This PR renames the top level package to src to avoid naming issues and
to represent the idea that the application contains other things like
CLI commands / background jobs. We considered naming it app, but that
would conflict with the app folder at the repo root, and it would also
create ambiguity with the flask app variable.