Add db-migrate to container path to meet infra template requirements (#176)
* Add db-migrate executable command to application container PATH
* Use consistent method for creating db connection between migrations
and application
* Remove code for make_connection_uri
* Remove support for offline migrations

## Context
These changes fulfills upcoming application requirements in
template-infra for using the database. The new requirement is that there
is a "db-migrate" command available in the container PATH.

The previous migrations code connects to the database using a different
method than what the application uses, which means that the migrations
code does not have the functionality of connecting to the DB using IAM
auth, which is a requirement for the infra template. In this change, we
make the migrations code in env.py use the existing PostgresDBClient
class and remove the extraneous make_connection_uri class.

As part of the change, we also remove support for offline migrations,
which is not needed in general. This simplifies the code paths in env.py
and reduces maintenance burden.