Upgrade Pydantic version (#209)
## Ticket

https://github.com/navapbc/template-application-flask/issues/208

## Changes

Upgrades to Pydantic v2

## Context for reviewers

https://docs.pydantic.dev/latest/migration/

Pydantic v2 is a fairly large rewrite of the library, although the
template only relies on it for env-var management. v2 is getting updated
consistently and is actively being worked on.

## Testing

`make format lint test` all work both running outside the Docker
container and within it.