Fix PYTHONPATH setting
In the container build instructions and docker compose, the source code
is mounted under an `app/` directory. So set `PYTHONPATH` to include
that in container environments.

It should not be in `local.env` as that is also used when running
natively, in which case such `PYTHONPATH` shenanigans should not be
necessary.

Resolves https://github.com/navapbc/template-application-flask/issues/246