Always use "app" in dockle accept file path (#253)
## Ticket

Resolves #252

## Changes

Fixed path for settings.py in DOCKLE_ACCEPT_FILES.

## Context for reviewers

Can be tested locally by exporting DOCKLE_ACCEPT_FILES before running
dockle. Without this fix, the settings.py file in the app causes that
check to fail. See the Dockerfile for where "app" is hardcoded in the
path:
https://github.com/navapbc/template-application-flask/blob/5d4d40aaa45d217d19e65814ff5f18cdff4b04fb/template/%7B%7Bapp_name%7D%7D/Dockerfile#L77

## Testing

Tested in a generated app.